Self register configuration resolver for web Part of #146027
